Implemented animating carousel
